+++ shop/src/main/java/ro/msg/learning/shop/repositories/SupplierRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import ro.msg.learning.shop.entities.Supplier;
 
 @Repository
-public interface SupplierRepository extends JpaRepository<Supplier,Long> {
+public interface SupplierRepository extends JpaRepository<Supplier,Integer> {
 
 }
